#5 final touches combat mode
