--- src/adventureGame/definitions.js
@@ -1,38 +1,0 @@
-export const OperandList = {
-	COMMAND: "command",
-	OBJECT: "object",
-	MOB: "mob",
-	ITEM: "item",
-	PATH: "path",
-	INVENTORY: "inventory",
-	KNOWN_SPELL: "known_spell",
-	RESPONSE: "response",
-	LOOK: "look"
-}
-
-export const SpellType = {
-	DAMAGE: "damage",
-	HEAL: "heal",
-	BUFF: "buff",
-	DEBUFF: "debuff"
-}
-
-export const MobStatus = {
-	ALIVE: "alive",
-	DEAD: "dead"
-}
-
-export const DamageType = {
-	ACID: "acid",
-	BLUDGEONING: "bludgeoning",
-	COLD: "cold",
-	FIRE: "fire",
-	FORCE: "force",
-	LIGHTNING: "lightning",
-	NECROTIC: "necrotic",
-	PIERCING: "piercing",
-	POISON: "poison",
-	PSYCHIC: "psychic",
-	RADIANT: "radiant",
-	SLASHING: "slashing"
-}
