+++ src/adventureGame/story/testStory.xml
@@ -175,6 +175,7 @@
 			<numberOfDice>1</numberOfDice>
 			<bonus>2</bonus>
 		</spell>
+
 	</spells>
 
 	<!-- Items -->
